homework_06/web_app/pyproject.toml
[tool.poetry]
name = "web_app"
version = "0.1.0"
description = ""
authors = ["Kerosin3 <keromvp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "<2.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.0.1"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
